+++ src/main/kotlin/com/freya02/botcommands/internal/parameters/resolvers/localization/LocalizationContextResolver.kt
@@ -1,0 +1,28 @@
+package com.freya02.botcommands.internal.parameters.resolvers.localization
+
+import com.freya02.botcommands.api.BContext
+import com.freya02.botcommands.api.localization.context.LocalizationContext
+import com.freya02.botcommands.api.parameters.ICustomResolver
+import com.freya02.botcommands.api.parameters.ParameterResolver
+import com.freya02.botcommands.internal.IExecutableInteractionInfo
+import com.freya02.botcommands.internal.localization.LocalizationContextImpl
+import com.freya02.botcommands.internal.throwInternal
+import net.dv8tion.jda.api.events.Event
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent
+import net.dv8tion.jda.api.interactions.Interaction
+
+internal class LocalizationContextResolver(private val baseContext: LocalizationContextImpl) :
+    ParameterResolver<LocalizationContextResolver, LocalizationContext>(LocalizationContext::class),
+    ICustomResolver<LocalizationContextResolver, LocalizationContext> {
+
+    override suspend fun resolveSuspend(context: BContext, executableInteractionInfo: IExecutableInteractionInfo, event: Event): LocalizationContext {
+        return when (event) {
+            is Interaction -> baseContext.withLocales(event.guildLocale, event.userLocale)
+            is MessageReceivedEvent -> when {
+                event.isFromGuild -> baseContext.withGuildLocale(event.guild.locale)
+                else -> baseContext
+            }
+            else -> throwInternal("Unsupported event type for localization contexts: ${event.javaClass.name}")
+        }
+    }
+}
